+++ packages/web3-utils/src/converters.ts
@@ -553,15 +553,20 @@ export const toWei = (number: Numbers, unit: EtherUnits): string => {
 	if (!denomination) {
 		throw new InvalidUnitError(unit);
 	}
-	if (typeof number === 'number'){
-		if (number  < 1e-15){
+	let parsedNumber = number;
+	if (typeof parsedNumber === 'number'){
+		if (parsedNumber  < 1e-15){
 			console.warn('Warning: The type `numbers` that are large or contain many decimals may cause loss of precision, it is recommended to use type `string` or `BigInt` when using conversion methods')
-			throw new InvalidNumberDecimalPrecisionLossError(number);
+			throw new InvalidNumberDecimalPrecisionLossError(number as number);
+		}
+		if (parsedNumber > 1e+20) {
+			console.warn('Warning: Using type `number` with values that are large or contain many decimals may cause loss of precision, it is recommended to use type `string` or `BigInt` when using conversion methods')
+
+			parsedNumber =  BigInt(parsedNumber);
 		}
 	}
 	
-	// create error if decimal place is over 20 digits
-	const parsedNumber = typeof number === 'number' ? number.toLocaleString('fullwide', {useGrouping: false, maximumFractionDigits: 20}) : number;
+	parsedNumber = typeof number === 'number' ? number.toLocaleString('fullwide', {useGrouping: false, maximumFractionDigits: 20}) : number;
 	// if value is decimal e.g. 24.56 extract `integer` and `fraction` part
 	// to avoid `fraction` to be null use `concat` with empty string
 	const [integer, fraction] = String(